+++ pkgs/dart_mcp/lib/src/client/client.dart
@@ -257,13 +257,19 @@ base class ServerConnection extends MCPBase {
   Future<CallToolResult> callTool(CallToolRequest request) =>
       sendRequest(CallToolRequest.methodName, request);
 
-  /// Lists all the resources from this server.
+  /// Lists all the [Resource]s from this server.
   Future<ListResourcesResult> listResources(ListResourcesRequest request) =>
       sendRequest(ListResourcesRequest.methodName, request);
 
-  /// Reads a [Resource] returned from the [ListResourcesResult].
+  /// Reads a [Resource] returned from the [ListResourcesResult] or matching
+  /// a [ResourceTemplate] from a [ListResourceTemplatesResult].
   Future<ReadResourceResult> readResource(ReadResourceRequest request) =>
       sendRequest(ReadResourceRequest.methodName, request);
+
+  /// Lists all the [ResourceTemplate]s from this server.
+  Future<ListResourceTemplatesResult> listResourceTemplates(
+    ListResourceTemplatesRequest request,
+  ) => sendRequest(ListResourceTemplatesRequest.methodName, request);
 
   /// Lists all the prompts from this server.
   Future<ListPromptsResult> listPrompts(ListPromptsRequest request) =>